requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-environ==0.11.2
django-solo==2.2.0
docutils==0.21.2
idna==3.7
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
python-dotenv==0.19.1
celery==5.1.2
redis==3.5.3

